requirements.txt
PyQt5
pyinstaller
pytesseract
pillow
pyautogui
pynput
mss
plyer 